timers: Add termdoro.sh
I'm not entirely sure why I didn't just integrate this with the break
cycle script I already wrote.

Reason for the temp file in the code is to make it easier on myself to
add nanowrimo word counts to my log without breaking focus until the
end.

Maybe merge the two scripts together later?

For now this works just fine.