git-utils: Add git-timelog.sh
The purpose of this script is to auto commit changes with a known start
and end date of authorship.

The reason why we don't use the diff's timestamp itself is when we want
to know the timeframe a change was written in. This is useful for things
like... nanowrimo.

Especially when you're cramped on time and need to focus on outputting
NOW now.